Init implementation of two templates (#1)
This PR covers initial implementation of two templates: `IoTDevice` and
`PeoplePayroll`. Data are pseudorandomly generated for most of the
fields. Some fields like names, etc. are picked from available list.

The usage should be in the clients where instead of static message we
will generated payload string through the generator.

---------

Signed-off-by: Jakub Stejskal <xstejs24@gmail.com>